+++ app/api/subject_area/slug/[slug]/route.ts
@@ -2,15 +2,19 @@ import { NextRequest, NextResponse } from 'next/server'
 import { createClient } from '@/utils/supabase/server'
 import { handleError } from '@/utils/errorHandler';
 
-export async function GET(context: { params: { slug: string } }) {
+export async function GET(
+  req: NextRequest, 
+  { params }: { params: Promise<{ slug: string }> }
+) {
   try {
-    const { slug } = await context.params;
+    const { slug } = await params;
 
     console.log('slug: ', slug);
 
     if (!slug) {
       return NextResponse.json({ error: 'Slug is required' }, { status: 400 });
     }
+    console.log("Extracted slug:", slug);
 
     const supabase = await createClient();
     const { data, error } = await supabase
