+++ app/api/topic_complex/slug/[slug]/route.ts
@@ -2,9 +2,12 @@ import { NextRequest, NextResponse } from 'next/server'
 import { createClient } from '@/utils/supabase/server'
 import { handleError } from '@/utils/errorHandler';
 
-export async function GET(context: { params: { slug: string } }) {
+export async function GET(
+  req: NextRequest, 
+  { params }: { params: Promise<{ slug: string }> }
+) {
   try {
-    const { slug } = await context.params;
+    const { slug } = await params;
 
     if (!slug) {
       return NextResponse.json({ error: 'Slug is required' }, { status: 400 });
