+++ app/api/topic_complex/subject/[id]/route.ts
@@ -2,13 +2,17 @@ import { NextRequest, NextResponse } from 'next/server';
 import { createClient } from '@/utils/supabase/server';
 import { handleError } from '@/utils/errorHandler';
 
-export async function GET(context: { params: { id: string } }) {
-    try {
-      const { id } = await context.params;
-  
-      if (!id) {
-        return NextResponse.json({ error: 'Id is required' }, { status: 400 });
-      }
+// export async function GET(_: NextRequest, context: { params: { id: string } }) {
+export async function GET(
+  req: NextRequest, 
+  { params }: { params: Promise<{ id: string }> }
+) {
+  try {
+    const { id } = await params;
+
+    if (!id) {
+      return NextResponse.json({ error: 'Id is required' }, { status: 400 });
+    }
 
     const supabase = await createClient();
     const { data, error } = await supabase
