+++ app/fachbereich/[slug]/[themenkomplex]/page.tsx
@@ -6,19 +6,14 @@ import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import { Badge } from "@/components/ui/badge"
 import { Breadcrumb } from "@/components/Breadcrumb"
-import { set } from "zod"
-import { Fachbereich } from "@/app/model/Fachbereich"
+// import { Fachbereich } from "@/app/model/Fachbereich"
 import { Question } from "../../../model/Question"
-import { Difficulty } from "@/app/model/Difficulty"
-import AddQuestionOverlay from "@/components/AddQuestionOverlay"
 import {
   Sheet,
   SheetContent,
   SheetHeader,
   SheetTitle,
   SheetDescription,
-  SheetFooter,
-  SheetClose,
 } from "@/components/ui/sheet"
 import {
   Dialog,
@@ -31,9 +26,11 @@ import {
 import AddQuestionForm from "@/components/AddQuestionOverlay"
 import EditQuestionForm from "@/components/EditQuestionForm"
 import { ApprenticeSelector } from "@/components/ApprenticeSelector"
-import { Themenkomplex } from "@/app/model/Themenkomplex"
+// import { Themenkomplex } from "@/app/model/Themenkomplex"
 
-export default function ThemenkomplexPage({ params }: { params: { slug: string; themenkomplex: string } }) {
+
+
+export default function ThemenkomplexPage({ params }: { params: Promise<{ slug: string; themenkomplex: string }> }) {
   const [expandedQuestions, setExpandedQuestions] = useState<number[]>([])
   const [isApprenticeSelectorOpen, setIsApprenticeSelectorOpen] = useState(false)
   const [questions, setQuestions] = useState<Question[]>([])
@@ -49,8 +46,8 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
   // const [currentSubject, setCurrentSubject] = useState<Fachbereich | null>(null)
   const [loading, setLoading] = useState(false);
 
-  const [subject, setSubject] = useState<Fachbereich | null>(null)
-  const [topic, setTopic] = useState<Themenkomplex | null>(null)
+  // const [subject, setSubject] = useState<Fachbereich | null>(null)
+  // const [topic, setTopic] = useState<Themenkomplex | null>(null)
 
   // init param
   useEffect(() => {
@@ -110,13 +107,15 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
   // once questions are fetched, fetch the subject and topic
   useEffect(() => {
 
+    console.log("fetch subject and topic", subjectSlug, topicSlug)
+
     if (!subjectSlug || !topicSlug) return;
 
     const getSubjectBySlug = async () => {
       try {
         const res = await fetch(`/api/subject_area/slug/${subjectSlug}`);
         const data = await res.json();
-        setSubject(data);
+        // setSubject(data);
         console.log("fetch subjects by slug", data);
       } catch (error) {
         console.error("Error fetching subject:", error);
@@ -127,7 +126,7 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
       try {
         const res = await fetch(`/api/topic_complex/slug/${topicSlug}`);
         const data = await res.json();
-        setTopic(data);
+        // setTopic(data);
         console.log("fetch topics by slug", data);
       } catch (error) {
         console.error("Error fetching topic:", error);
@@ -137,9 +136,9 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
     getSubjectBySlug();
     getTopicBySlug();
 
-  }, [questions])
+  }, [questions, subjectSlug, topicSlug])
 
-  const handleEditQuestion = async (data: any) => {
+  const handleEditQuestion = async (data: { id: number; question: string; answer: string; difficulty: number; tags: string[] }) => {
     try {
       setLoading(true)
       console.log("data in handleEditQuestion", data);
@@ -168,7 +167,7 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
     }
   }
 
-  const handleAddQuestion = async (data: any) => {
+  const handleAddQuestion = async (data: { question: string; answer: string; difficulty: number; tags: string[] }) => {
     try {
       setLoading(true)
       console.log("data here:", data)
@@ -179,7 +178,7 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
           ...data, 
           topic_id: topicId, 
           difficulty_id: data.difficulty, 
-          tags: data.tags.split(",").map((tag: string) => tag.trim())
+          tags: data.tags
         }),
       })
 
@@ -428,7 +427,7 @@ export default function ThemenkomplexPage({ params }: { params: { slug: string; 
       <Dialog open={isDeleteOpen} onOpenChange={setIsDeleteOpen}>
         <DialogContent>
           <DialogHeader>
-            <DialogTitle>Delete Question "{currentQuestion?.question}"</DialogTitle>
+            <DialogTitle>Delete Question &quot;{currentQuestion?.question}&quot;</DialogTitle>
             <DialogDescription>
               Are you sure you want to delete this question? This action cannot be undone.
             </DialogDescription>
