+++ app/questions/page.tsx
@@ -227,7 +227,11 @@ export default function QuestionsPage() {
         </div>
 
         <div className="space-y-4">
-          {filteredQuestions.map((question) => (
+          { loading ? (
+            <p>Loading...</p>
+          ) : questionsDb.length === 0 ? (
+            <p>No questions found.</p>
+          ) : filteredQuestions.map((question) => (
             <Card key={question.id} className="relative">
               <CardHeader>
                 <CardTitle className="text-lg pr-20">{question.question}</CardTitle>
