+++ package.json
@@ -27,7 +27,7 @@
     "clsx": "^2.1.1",
     "jspdf-autotable": "^5.0.2",
     "lucide-react": "^0.475.0",
-    "next": "15.1.7",
+    "next": "^15.2.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
